Fully implement inserting values into the tree
The `SWBinarySearchTree` serves as a wrapper and public interface to the tree nodes. For insertion it
delegates to the nodes which decide in turn where the value should go. Types that can be inserted are
limited to those that have a defined ordering and can be compared through the usual boolean 
comparision operations i.e. greater than, less than etc. 

The `SWNodeTest` handles value insertion well. In future `SWBinarySearchTreeTest` should remove
tests covering where values are inserted and instead keep track of how many values are held.

Operations to add include searching for values, deleting values, returning the tree size. Any
work to make the tree self balancing will be considered after the core features are implemented.