Re-implement equality testing
Previously equality was tested by an overridden `#=` method. This
is not enough for equality. `#hash` must also be overridden as it
is used for sets and dictionaries. This could lead to very hard to find
bugs.

`#hash` must always return the same value for the lifetime of an
object. A common approach is to take the XOR of the hash values
of immutable instance variables. Neither `SWBinarySearchTree`
or `SWNode` have  instance variables that will remain constant
over its lifetime so this was not an option. Instead methods have 
been implemented for both that test if the contents of two trees 
are identical. An equality check with `#=` will still test object 
identity.

Additionally fixed a bug that would allow certain trees with differing
contents to be counted as equal. The error was to use a message
cascade to test multiple boolean conditions.